add Flyway

+++ src/main/java/com/example/sklad/model/entities/Company.java
@@ -21,7 +21,7 @@ public class Company {
 
     private String name;
 
-    private int inn;
+    private long inn;
 
     private int kpp;
